move unit test configuration so it's included in test configuration
